+++ constants.py
@@ -1,5 +1,5 @@
 APP_NAME = "KönyvTárnok"
-APP_VERSION = "2.1.1"
+APP_VERSION = "3.0.0"
 APP_STAGE = ""
 APP_TITLE = APP_NAME